fix a regression in FID computation introduced in https://github.com/toshas/torch-fidelity/issues/46 https://github.com/toshas/torch-fidelity/pull/50 through differences in behavior of torch and np (.linalg.eigvals) and resulting in rare NaN values introduce additional precision testing for the new implementation to ensure consistency with torch-fidelity<=0.3.0

+++ torch_fidelity/metric_fid.py
@@ -31,7 +31,7 @@ def fid_statistics_to_metric(stat_1, stat_2, verbose):
     assert sigma1.ndim == 2 and sigma1.shape == sigma2.shape and sigma1.dtype == sigma2.dtype
 
     diff = mu1 - mu2
-    tr_covmean = np.sum(np.sqrt(np.linalg.eigvals(sigma1.dot(sigma2))).real)
+    tr_covmean = np.sum(np.sqrt(np.linalg.eigvals(sigma1.dot(sigma2)).astype('complex128')).real)
     fid = float(diff.dot(diff) + np.trace(sigma1) + np.trace(sigma2) - 2 * tr_covmean)
 
     out = {
